Combine database scripts, make them atomic, and change from delete/repopulate to upsert strategy
