+++ update_data.py
@@ -16,6 +16,7 @@ import psycopg2
 import sys
 import os
 import geocoder
+import requests
 from dotenv import load_dotenv
 
 
@@ -36,19 +37,162 @@ def get_env_var(var_name, fallback=""):
     return value
 
 
-def delete_data(cursor):
-    cursor.execute('DELETE FROM geolocations')
-    cursor.execute('DELETE FROM location_aliases')
-    logging.info('Deleted existing tables')
+def get_people(token):
+    people = []
+
+    session = requests.session()
+    session.headers.update({'Authorization': f'Bearer {token}'})
+    url = 'https://www.recurse.com/api/v1/profiles?limit={limit}&offset={offset}'
+    limit = 50
+    offset = 0
+
+    while True:
+        r = session.get(url.format(limit=limit, offset=offset))
+        if r.status_code != requests.codes.ok:
+            r.raise_for_status()
+        page = r.json()
+        if page == []:
+            break
+        people.extend(page)
+        offset += limit
+
+    return people
 
 
-def replace_data(database_url):
+def insert_people_data(cursor, people):
+    processed_batches = set()
+    processed_locations = set()
+
+    for person in people:
+        person_id = person.get('id')
+        image_path = person.get('image_path')
+        image_url = image_path if ("no_photo" not in image_path) else None
+
+        first_name = person.get('first_name')
+        middle_name = person.get('middle_name')
+        last_name = person.get('last_name')
+
+        logging.debug("Person #{}: {} {} {}".format(
+            person_id,
+            first_name,
+            middle_name,
+            last_name
+        ))
+
+        cursor.execute("INSERT INTO people" +
+                       " (person_id, first_name, middle_name," +
+                       "  last_name, image_url)" +
+                       " VALUES (%s, %s, %s, %s, %s)" +
+                       " ON CONFLICT (person_id) DO UPDATE SET" +
+                       " first_name = %s," +
+                       " middle_name = %s," +
+                       " last_name = %s," +
+                       " image_url = %s" +
+                       " WHERE people.person_id = %s",
+                       [person_id,
+                        first_name,
+                        middle_name,
+                        last_name,
+                        image_url,
+                        first_name,
+                        middle_name,
+                        last_name,
+                        image_url,
+                        person_id
+                        ]
+                       )
+
+        location = person.get('current_location')
+        if (location):
+            location_id = location.get('id')
+
+            if (location_id not in processed_locations):
+                processed_locations.add(location_id)
+                logging.debug("Location #{}: {} ({})".format(
+                    location_id,
+                    location.get('name'),
+                    location.get('short_name')
+                ))
+                cursor.execute("INSERT INTO locations" +
+                               " (location_id, name, short_name)" +
+                               " VALUES (%s, %s, %s)" +
+                               " ON CONFLICT (location_id) DO NOTHING",
+                               [location_id,
+                                location.get('name'),
+                                location.get('short_name')
+                                ]
+                               )
+
+            cursor.execute("INSERT INTO location_affiliations" +
+                           " (person_id, location_id, affiliation_type)" +
+                           " VALUES (%s, %s, %s)" +
+                           " ON CONFLICT (person_id, affiliation_type)" +
+                           " DO UPDATE SET" +
+                           " location_id = %s" +
+                           " WHERE location_affiliations.person_id = %s",
+                           [person_id,
+                            location_id,
+                            "current_location",
+                            location_id,
+                            person_id
+                            ]
+                           )
+
+        for stint in person['stints']:
+            batch_id = None
+            batch = stint.get('batch')
+
+            if (batch):
+                batch_id = batch.get('id')
+
+                if (batch_id not in processed_batches):
+                    logging.debug("  Batch {}, \"{}\" ({})".format(
+                        batch_id,
+                        batch.get('name'),
+                        batch.get('short_name')
+                    ))
+                    cursor.execute("INSERT INTO batches " +
+                                   " (batch_id, name, short_name)" +
+                                   " VALUES (%s, %s, %s)" +
+                                   " ON CONFLICT (batch_id) DO NOTHING",
+                                   [batch_id,
+                                    batch.get('name'),
+                                    batch.get('short_name')
+                                    ]
+                                   )
+                    processed_batches.add(batch_id)
+
+            logging.debug("  Stint: {}, Batch: {}, {} - {}".format(
+                stint.get('type'),
+                batch_id,
+                stint.get('start_date'),
+                stint.get('end_date')
+            ))
+            cursor.execute("INSERT INTO stints" +
+                           " (person_id, batch_id, stint_type," +
+                           "  start_date, end_date, title)" +
+                           " VALUES (%s, %s, %s, %s, %s, %s)" +
+                           " ON CONFLICT (person_id, start_date) DO NOTHING",
+                           [person_id,
+                            batch_id,
+                            stint.get('type'),
+                            stint.get('start_date'),
+                            stint.get('end_date'),
+                            stint.get('title'),
+                            ]
+                           )
+
+    logging.info('Inserted %s people', len(people))
+    logging.info('Inserted %s batches', len(processed_batches))
+    logging.info('Inserted %s locations', len(processed_locations))
+
+
+def process_rc_data(database_url, people):
     connection = psycopg2.connect(database_url)
     cursor = connection.cursor()
 
-    delete_data(cursor)
-    insert_data(cursor)
-    reconcile_duplicates(cursor)
+    insert_people_data(cursor, people)
+    add_geolocation(cursor)
 
     connection.commit()
     cursor.close()
@@ -56,17 +200,22 @@ def replace_data(database_url):
     logging.info('Completed database update')
 
 
-def insert_data(cursor):
-    logging.info('Beginning database update')
+def add_geolocation(cursor):
     locations = get_locations_from_db(cursor)
-    logging.info('Retrieved %s locations', len(locations))
-    logging.info('Adding geolocation data... (this will take a few minutes)')
+    no_geo_count = len(locations)
 
-    for location in locations:
-        geo = lookup_geodata(cursor, location)
-        insert_geo_data(cursor, geo)
+    logging.info('Retrieved %s locations without geo data', no_geo_count)
 
-    logging.info('Inserted %s locations', len(locations))
+    if (no_geo_count > 0):
+        logging.info(
+            'Adding geolocation data... (this will take a few minutes)')
+
+        for location in locations:
+            geo = lookup_geodata(cursor, location)
+            insert_geo_data(cursor, geo)
+
+        logging.info('Inserted %s locations', no_geo_count)
+        reconcile_duplicates(cursor)
 
 
 def lookup_geodata(cursor, location):
@@ -83,7 +232,12 @@ def get_locations_from_db(cursor):
     cursor.execute("""SELECT
                         location_id,
                         name
-                      FROM locations
+                      FROM locations l
+                      WHERE NOT EXISTS (
+                        SELECT FROM geolocations g 
+                        WHERE
+                         g.location_id = l.location_id
+                      )
                       ORDER BY location_id""")
     return [{
         'location_id': x[0],
@@ -92,6 +246,7 @@ def get_locations_from_db(cursor):
 
 
 geonames_username = get_env_var('GEONAMES_USERNAME')
+
 
 def geonames_query(location):
     time.sleep(2.5)  # Slow down requests to avoid timeout (< 1/sec)
@@ -107,6 +262,7 @@ def geonames_query(location):
     feature_classes = ['A', 'P']
 
     return geocoder.geonames(query, key=geonames_username, featureClass=feature_classes, maxRows=1).json["raw"]
+
 
 def get_state_name(state_code):
     usStates = {}
@@ -141,9 +297,9 @@ def parse_location(location):
 
 def add_geonames_result(parsed_location, geonames_result):
     logging.debug("Combining geo results. Parsed: {}\n GeoNames: {}".format(
-            parsed_location,
-            geonames_result
-        ))
+        parsed_location,
+        geonames_result
+    ))
 
     return {
         'location_id': parsed_location['location_id'],
@@ -174,7 +330,8 @@ def insert_geo_data(cursor, location):
                    " (location_id, name, type," +
                    " subdivision_derived, subdivision_code," +
                    " country_name, country_code, lat, lng)" +
-                   " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s)",
+                   " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s)" +
+                   " ON CONFLICT (location_id) DO NOTHING",
                    [location.get('location_id'),
                     location.get('name'),
                     location.get('type'),
@@ -198,7 +355,8 @@ def insert_alias(cursor, location, preferred_location):
 
     cursor.execute("INSERT INTO location_aliases" +
                    " (location_id, preferred_location_id)" +
-                   " VALUES (%s, %s)",
+                   " VALUES (%s, %s)" +
+                   " ON CONFLICT (location_id) DO NOTHING",
                    [location.get('location_id'),
                     preferred_location.get('location_id')])
 
@@ -230,9 +388,6 @@ def reconcile_duplicates(cursor):
 
             # Update all people pointing to old location with new in location_affiliations
             update_location_affiliations(cursor, loc, preferred_loc)
-
-            # Delete duplicate from geolocations
-            delete_geolocation(cursor, loc)
 
     logging.info('Duplicate locations removed')
 
@@ -276,22 +431,15 @@ def update_location_affiliations(cursor, old_location, new_location):
                    [new_location["location_id"], old_location["location_id"]])
 
 
-def delete_geolocation(cursor, location):
-    logging.info("Delete Geolocation for #{} {}".format(
-        location.get('location_id'),
-        location.get('name')
-    ))
-
-    """Removes the row with the given location_id from the geolocation table."""
-    cursor.execute("""DELETE 
-                      FROM geolocations
-                      WHERE location_id = %s""", [location["location_id"]])
-
-
 if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO)
     database_url = get_env_var('DATABASE_URL')
+    token = get_env_var('RC_API_ACCESS_TOKEN')
 
-    logging.info('Starting World Map geolocation update...')
-    replace_data(database_url)
+    logging.info('Starting World Map database update')
+    logging.info(
+        'Pulling RC API profile data (this may take a few seconds)...')
+    people = get_people(token)
+    logging.info('Found %s people', len(people))
+    process_rc_data(database_url, people)
     logging.info('Geolocation completed.')
